+++ _userspace/static-pages/404.php
@@ -1,0 +1,6 @@
+<?php
+$req = $_SERVER['REQUEST_URI'];
+$req = preg_replace('#static-page/#','',$req);
+header("Location: $req");
+die();
+?>